Refactor database utils into class
